backend/requirements.txt
flask==2.2.5
flask-cors==3.0.10
flask-socketio==5.3.6
python-dotenv==1.0.0
eventlet==0.33.3
requests==2.31.0 